Fix bug - Use dynamic box size
Previously, if someone would change the PLOT_SIZE constant to something like 600dp that wouldn't fit on a smaller device - we wouldn't get our guaranteed square plot on the screen.

So now i'm storing the "current box size" (after it's aspect ratio and preferred size have been set) so that I can take into account any adjustments that need to be made in order to plot the dot!